полностью доработал heredoc а так же его возвращаемые статусы

+++ sources/execution/redirections.c
@@ -6,7 +6,7 @@
 /*   By: ggevorgi <sp1tak.gg@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/27 12:41:01 by mikayel           #+#    #+#             */
-/*   Updated: 2025/06/02 13:05:41 by ggevorgi         ###   ########.fr       */
+/*   Updated: 2025/06/03 12:24:22 by ggevorgi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,6 @@ void    exit_error(char *name)
     char    *msg;
     
     msg = ft_strjoin("minishell: ", name);
-	printf("chmo1\n");
     perror(msg);
     free(msg);
     exit(EXIT_FAILURE);
@@ -27,14 +26,17 @@ void	open_dup_fd(t_redirection *redir, int redir_fd, int flags, mode_t mode)
 {
 	int	fd;
 	
-	fd = open(redir->target, flags, mode);
-	if (fd == -1)
-		exit_error(redir->target);
-	if (redir->type == REDIR_HEREDOC)
+	if (!redir->target)
 	{
+		signal_status = 130;
+		exit(130);
+	}
+	fd = open(redir->target, flags, mode);
+	if (fd == -1)	
+		exit_error(redir->target);
+	if (redir->type == REDIR_HEREDOC)		
 		if (unlink(redir->target) == -1)
 			exit_error(redir->target);
-	}
 	if (dup2(fd, redir_fd) == -1)
 		exit_error(redir->target);
 	if (close(fd) == -1)
